feat(desktop-app): time range picker, refresh picker, and UI improvements
- Move TimeRangePicker and RefreshPicker to @everr/ui as controlled
  components, reuse them in both webapp and desktop app
- Add from/to time range params to get_runs_list Tauri command
- Use React Query refetchInterval with refetchIntervalInBackground:false
  instead of manual setInterval for auto-refresh
- Make workflow name clickable to open run in browser, move copy
  auto-fix button to its own column
- Add table-fixed layout with truncation for narrow windows
- Use Empty component for the no-runs state
- Guard /developer route with redirect in production builds
- Fix window focus: set ActivationPolicy::Regular before showing window
- Reduce notification window right margin from 56px to 16px
