+++ packages/desktop-app/src-tauri/src/lib.rs
@@ -54,7 +54,7 @@ const NOTIFICATION_EXIT_EVENT: &str = "everr://notification-exit";
 const NOTIFICATION_WINDOW_LABEL: &str = "notification";
 const NOTIFICATION_WINDOW_WIDTH: f64 = 420.0;
 const NOTIFICATION_WINDOW_HEIGHT: f64 = 124.0;
-const NOTIFICATION_WINDOW_MARGIN: f64 = 56.0;
+const NOTIFICATION_WINDOW_MARGIN: f64 = 16.0;
 const NOTIFICATION_WINDOW_INSET: f64 = 12.0;
 const TRAY_ICON_ID: &str = "everr-app";
 const SETTINGS_MENU_ID: &str = "settings";
